factor list_objects_with_sets out of NarrativeManager
