+++ lib/NarrativeService/NarrativeManager.py
@@ -21,131 +21,13 @@ class NarrativeManager:
 
     DEBUG = False
 
-    def __init__(self, config, ctx, set_api_client, data_palette_client):
+    def __init__(self, config, user_id, set_api_client, data_palette_client, workspace_client):
         self.narrativeMethodStoreURL = config["narrative-method-store"]
-        self.set_api_cache = set_api_client  # DynamicServiceCache type
+        self.set_api_client = set_api_client                     # DynamicServiceCache type
         self.data_palette_client = data_palette_client          # DynamicServiceCache type
-        self.token = ctx["token"]
-        self.user_id = ctx["user_id"]
-        self.ws = Workspace(config["workspace-url"], token=self.token)
+        self.user_id = user_id
+        self.ws = workspace_client
         self.intro_md_file = config["intro-markdown-file"]
-
-    def list_objects_with_sets(self, ws_id=None, ws_name=None, workspaces=None,
-                               types=None, include_metadata=0, include_data_palettes=0):
-        if not workspaces:
-            if not ws_id and not ws_name:
-                raise ValueError("One and only one of 'ws_id', 'ws_name', 'workspaces' " +
-                                 "parameters should be set")
-            workspaces = [self._get_workspace_name_or_id(ws_id, ws_name)]
-        return self._list_objects_with_sets(workspaces, types, include_metadata, include_data_palettes)
-
-    def _list_objects_with_sets(self, workspaces, types, include_metadata, include_data_palettes):
-        type_map = None
-        if types is not None:
-            type_map = {key: True for key in types}
-
-        processed_refs = {}
-        data = []
-        if self.DEBUG:
-            print("NarrativeManager._list_objects_with_sets: processing sets")
-        t1 = time.time()
-        set_ret = self.set_api_cache.call_method("list_sets",
-                                                 [{'workspaces': workspaces,
-                                                   'include_set_item_info': 1,
-                                                   'include_metadata': include_metadata}],
-                                                 self.token)
-        sets = set_ret['sets']
-        for set_info in sets:
-            # Process
-            target_set_items = []
-            for set_item in set_info['items']:
-                target_set_items.append(set_item['info'])
-            if self._check_info_type(set_info['info'], type_map):
-                data_item = {'object_info': set_info['info'],
-                             'set_items': {'set_items_info': target_set_items}}
-                data.append(data_item)
-                processed_refs[set_info['ref']] = data_item
-        if self.DEBUG:
-            print("    (time=" + str(time.time() - t1) + ")")
-
-        if self.DEBUG:
-            print("NarrativeManager._list_objects_with_sets: loading ws_info")
-        t2 = time.time()
-        ws_info_list = []
-        # for ws in workspaces:
-        if len(workspaces) == 1:
-            ws = workspaces[0]
-            ws_id = None
-            ws_name = None
-            if str(ws).isdigit():
-                ws_id = int(ws)
-            else:
-                ws_name = str(ws)
-            ws_info_list.append(self.ws.get_workspace_info({"id": ws_id, "workspace": ws_name}))
-        else:
-            ws_map = {key: True for key in workspaces}
-            for ws_info in self.ws.list_workspace_info({'perm': 'r'}):
-                if ws_info[1] in ws_map or str(ws_info[0]) in ws_map:
-                    ws_info_list.append(ws_info)
-        if self.DEBUG:
-            print("    (time=" + str(time.time() - t2) + ")")
-
-        if self.DEBUG:
-            print("NarrativeManager._list_objects_with_sets: loading workspace objects")
-        t3 = time.time()
-        for info in WorkspaceListObjectsIterator(self.ws,
-                                                 ws_info_list=ws_info_list,
-                                                 list_objects_params={
-                                                     'includeMetadata': include_metadata
-                                                 }):
-            item_ref = str(info[6]) + '/' + str(info[0]) + '/' + str(info[4])
-            if item_ref not in processed_refs and self._check_info_type(info, type_map):
-                data_item = {'object_info': info}
-                data.append(data_item)
-                processed_refs[item_ref] = data_item
-        if self.DEBUG:
-            print("    (time=" + str(time.time() - t3) + ")")
-
-        return_data = {
-            "data": data
-        }
-
-        if include_data_palettes == 1:
-            if self.DEBUG:
-                print("NarrativeManager._list_objects_with_sets: processing DataPalettes")
-            t5 = time.time()
-            dp_ret = self.data_palette_client.call_method(
-                "list_data",
-                [{'workspaces': workspaces, 'include_metadata': include_metadata}],
-                self.token
-            )
-            for item in dp_ret['data']:
-                ref = item['ref']
-                if self._check_info_type(item['info'], type_map):
-                    data_item = None
-                    if ref in processed_refs:
-                        data_item = processed_refs[ref]
-                    else:
-                        data_item = {'object_info': item['info']}
-                        processed_refs[ref] = data_item
-                        data.append(data_item)
-                    dp_info = {}
-                    if 'dp_ref' in item:
-                        dp_info['ref'] = item['dp_ref']
-                    if 'dp_refs' in item:
-                        dp_info['refs'] = item['dp_refs']
-                    data_item['dp_info'] = dp_info
-            return_data["data_palette_refs"] = dp_ret['data_palette_refs']
-            if self.DEBUG:
-                print("    (time=" + str(time.time() - t5) + ")")
-
-        return return_data
-
-    def _check_info_type(self, info, type_map):
-        if type_map is None:
-            return True
-        obj_type = info[2].split('-')[0]
-        return type_map.get(obj_type, False)
 
     def copy_narrative(self, newName, workspaceRef, workspaceId):
         time_ms = int(round(time.time() * 1000))
@@ -312,7 +194,7 @@ class NarrativeManager:
                 appSpecIds.append(cell['app'])
             elif 'method' in cell:
                 methodSpecIds.append(cell['method'])
-        nms = NarrativeMethodStore(self.narrativeMethodStoreURL, token=self.token)
+        nms = NarrativeMethodStore(self.narrativeMethodStoreURL)
         if len(appSpecIds) > 0:
             appSpecs = nms.get_app_spec({'ids': appSpecIds})
             for spec in appSpecs:
@@ -468,12 +350,6 @@ class NarrativeManager:
             pass  # it's boolean/int/float/None
         return obj
 
-    def _get_workspace_name_or_id(self, ws_id, ws_name):
-        ret = ws_name
-        if not ret:
-            ret = str(ws_id)
-        return ret
-
     def copy_object(self, ref, target_ws_id, target_ws_name, target_name, src_info):
         """
         Copies an object from one workspace to another.
@@ -497,14 +373,3 @@ class NarrativeManager:
         obj_info = ServiceUtils.object_info_to_object(obj_info_tuple)
         return {'info': obj_info}
 
-    def list_available_types(self, workspaces):
-        data = self.list_objects_with_sets(workspaces=workspaces)['data']
-        type_stat = {}
-        for item in data:
-            info = item['object_info']
-            obj_type = info[2].split('-')[0]
-            if obj_type in type_stat:
-                type_stat[obj_type] += 1
-            else:
-                type_stat[obj_type] = 1
-        return {'type_stat': type_stat}
