+++ lib/NarrativeService/authclient.py
@@ -45,7 +45,7 @@ class TokenCache(object):
             self._cache[token] = [user, _time.time()]
             if len(self._cache) > self._maxsize:
                 for i, (t, _) in enumerate(sorted(self._cache.items(),
-                                                  key=lambda (_, v): v[1])):
+                                                  key=lambda v: v[1])):
                     if i <= self._halfmax:
                         del self._cache[t]
                     else:
